# Domain/Constants/DonationEnums.cs
namespace ClubManagement.Domain.Constants
{
    public enum DonationType
    {
        Monthly = 1,
        General = 2,
        Event   = 3
    }

    public enum PaymentMethod
    {
        Cash  = 1,
        Bkash = 2,
        Bank  = 3
    }
}