# client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ClubManagement",
    "version": "1.0"
  },
  "paths": {
...
    "/api/donations": {
      "get": {
        "tags": [
          "Donation"
        ],
        "parameters": [
          {
            "name": "MemberId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DonationType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DonationType"
            }
          },
          {
            "name": "PaymentMethod",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PaymentMethod"
            }
          },
          {
            "name": "FromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ToDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Donation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDonationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/donations/{id}": {
      "get": {
        "tags": [
          "Donation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/members/{memberId}/donations": {
      "get": {
        "tags": [
          "Donation"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateDonationDto": {
        "required": [
          "amount",
          "donationDate",
          "donationType",
          "memberId",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "maximum": 1.7976931348623157E+308,
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "donationType": {
            "$ref": "#/components/schemas/DonationType"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "referenceNumber": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "donationDate": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DonationType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PaymentMethod": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  },
...
}